moved update post component to be under post feature instead of as a shared component
